feat: add auth to user

+++ src/modules/users/users.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -11,15 +12,20 @@ import { UsersRepository } from './repository/users.repository';
 export class UsersService {
   constructor(private repository: UsersRepository) {}
   async create(data: CreateUserDto) {
-    const usernameExists = await this.repository.findByUsername(data.username);
+    const { username, email, phone } = data;
+    const user = await this.repository.findByUser(username, email, phone);
 
-    if (usernameExists)
-      throw new ConflictException({ username: 'Username already exists.' });
+    if (user) {
+      const message = [];
 
-    const emailExists = await this.repository.findByEmail(data.email);
+      if (user.username == username) message.push('Username already exists.');
 
-    if (emailExists)
-      throw new ConflictException({ email: 'Email already exists' });
+      if (user.email == email) message.push('Email already exists.');
+
+      if (user.phone == phone) message.push('Phone already exists.');
+
+      throw new NotFoundException({ message });
+    }
 
     return await this.repository.create(data);
   }
@@ -36,9 +42,8 @@ export class UsersService {
       ? await this.repository.logByEmail(credential)
       : await this.repository.logByUsername(credential);
 
-    if (!user) {
-      throw new NotFoundException('User does not exists.');
-    }
+    if (!user)
+      throw new NotFoundException({ message: 'User does not exists.' });
 
     return user;
   }
@@ -46,29 +51,38 @@ export class UsersService {
   async findOne(id: number) {
     const user = await this.repository.find(id);
 
-    if (!user) {
-      throw new NotFoundException('User does not exists.');
-    }
+    if (!user)
+      throw new NotFoundException({ message: 'User does not exists.' });
 
     return user;
   }
 
-  async update(id: number, data: UpdateUserDto) {
+  async update(id: number, data: UpdateUserDto, userId: number) {
     const user = await this.repository.update(id, data);
 
-    if (!user) {
-      throw new NotFoundException('User does not exists.');
-    }
+    if (!user)
+      throw new NotFoundException({ message: 'User does not exists.' });
+
+    if (user.id != userId)
+      throw new ForbiddenException({
+        message:
+          'You do not have sufficient permissions to perform this action.',
+      });
 
     return user;
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
     const user = await this.repository.find(id);
 
-    if (!user) {
-      throw new NotFoundException('User does not exists.');
-    }
+    if (!user)
+      throw new NotFoundException({ message: 'User does not exists.' });
+
+    if (user.id != userId)
+      throw new ForbiddenException({
+        message:
+          'You do not have sufficient permissions to perform this action.',
+      });
 
     return await this.repository.delete(id);
   }
